+++ apps/next/components/Information/Information.tsx
@@ -1,35 +1,24 @@
 'use client'
 
-import { Button, Text } from '@midas-ds/components'
-import { Info } from 'lucide-react'
-import { useState } from 'react'
-import styles from './Information.module.css'
-import { Panel } from '@midas-ds/layout'
+import { Button } from '@midas-ds/components'
+import { useDismissPanels } from '@midas-ds/layout'
 
 export const Information = () => {
-  const [isDismissPanelOpen, setIsDismissPanelOpen] = useState(false)
+  const { addPanel } = useDismissPanels()
 
   return (
-    <Text className={styles.information}>
-      This app tests components
-      <Button
-        className={styles.icon}
-        variant='icon'
-        size='medium'
-        onPress={() => setIsDismissPanelOpen(prev => !prev)}
-      >
-        <Info size={20} />
-      </Button>
-      <Panel
-        variant='dismiss'
-        title='Details'
-        aria-label='details panel'
-        data-debug='Panel (dismiss)'
-        isOpen={isDismissPanelOpen}
-        onOpenChange={setIsDismissPanelOpen}
-      >
-        More information
-      </Panel>
-    </Text>
+    <Button
+      variant='secondary'
+      onPress={() =>
+        addPanel({
+          id: 'details',
+          title: 'Details',
+          'aria-label': 'details panel',
+          children: 'More information',
+        })
+      }
+    >
+      Open panel
+    </Button>
   )
 }