+++ packages/layout/src/layout/layout-content/LayoutContent.tsx
@@ -1,0 +1,14 @@
+import { DetailedHTMLProps, HTMLAttributes } from 'react'
+import styles from './LayoutContent.module.css'
+
+export type LayoutContentProps = DetailedHTMLProps<
+  HTMLAttributes<HTMLDivElement>,
+  HTMLDivElement
+>
+
+export const LayoutContent = ({ className, ...rest }: LayoutContentProps) => (
+  <div
+    className={`${className ?? ''} ${styles.layoutContent}`}
+    {...rest}
+  />
+)