+++ packages/layout/src/panel/panel-body/PanelBody.tsx
@@ -1,15 +1,15 @@
-import { DetailedHTMLProps, HTMLAttributes } from 'react'
+import { forwardRef, HTMLAttributes } from 'react'
 import { clsx } from '@midas-ds/components'
 import styles from './PanelBody.module.css'
 
-export type PanelBodyProps = DetailedHTMLProps<
-  HTMLAttributes<HTMLElement>,
-  HTMLElement
->
+export type PanelBodyProps = HTMLAttributes<HTMLDivElement>
 
-export const PanelBody = ({ className, ...rest }: PanelBodyProps) => (
-  <aside
-    className={clsx(className, styles.panelBody)}
-    {...rest}
-  />
+export const PanelBody = forwardRef<HTMLDivElement, PanelBodyProps>(
+  ({ className, ...rest }, ref) => (
+    <div
+      ref={ref}
+      className={clsx(className, styles.panelBody)}
+      {...rest}
+    />
+  ),
 )