+++ packages/layout/CHANGELOG.md
@@ -1,3 +1,83 @@
+## 0.0.22 (2026-03-24)
+
+### 🧱 Updated Dependencies
+
+- Updated components to 17.10.1
+
+## 0.0.21 (2026-03-23)
+
+### 🩹 Fixes
+
+- **layout, panel:** move transition css to collapse-panel ([5f45327752](https://github.com/migrationsverket/midas/commit/5f45327752))
+- **layout, panel:** remove conflicting css of same specificity ([09410df703](https://github.com/migrationsverket/midas/commit/09410df703))
+- **layout:** externalize components dependency ([3647b3fb11](https://github.com/migrationsverket/midas/commit/3647b3fb11))
+
+## 0.0.20 (2026-03-23)
+
+### 🩹 Fixes
+
+- **layout, navigation:** decouple navigation-link from link component ([d33aa75a22](https://github.com/migrationsverket/midas/commit/d33aa75a22))
+
+## 0.0.19 (2026-03-20)
+
+### 🩹 Fixes
+
+- **layout, panel:** rename context provider ([6996c35755](https://github.com/migrationsverket/midas/commit/6996c35755))
+- **layout:** replace setPanels with addPanel/closePanel API and add bring-to-front behavior ([a151c18a8c](https://github.com/migrationsverket/midas/commit/a151c18a8c))
+- **layout, panel:** add controlled state support ([d7d5777852](https://github.com/migrationsverket/midas/commit/d7d5777852))
+- **layout, panel:** unshift and aria hide nested panels ([21c7b4ba91](https://github.com/migrationsverket/midas/commit/21c7b4ba91))
+- **layout, panel:** change api to support multiple panels ([9987fe9ebe](https://github.com/migrationsverket/midas/commit/9987fe9ebe))
+- **layout, panel:** make dismiss panel globally configurable ([eadc5621b2](https://github.com/migrationsverket/midas/commit/eadc5621b2))
+
+## 0.0.18 (2026-03-19)
+
+### 🧱 Updated Dependencies
+
+- Updated components to 17.10.0
+
+## 0.0.17 (2026-03-19)
+
+### 🧱 Updated Dependencies
+
+- Updated components to 17.9.0
+
+## 0.0.16 (2026-03-18)
+
+### 🧱 Updated Dependencies
+
+- Updated components to 17.8.3
+
+## 0.0.15 (2026-03-17)
+
+This was a version bump only for layout to align it with other projects, there were no code changes.
+
+## 0.0.14 (2026-03-16)
+
+### 🧱 Updated Dependencies
+
+- Updated components to 17.8.2
+
+## 0.0.13 (2026-03-13)
+
+### 🩹 Fixes
+
+- **layout,panel:** suppress animation for prefers-reduced-motion ([b761e5d4f7](https://github.com/migrationsverket/midas/commit/b761e5d4f7))
+- **layout:** add overflow hidden to content to prevent scrollbar during animation ([7078674395](https://github.com/migrationsverket/midas/commit/7078674395))
+- **layout,panel:** add forwardRef to PanelBody for React 18 compat ([a0b06528a6](https://github.com/migrationsverket/midas/commit/a0b06528a6))
+- **layout, panel:** add animations to dismiss-panel ([b0e8733489](https://github.com/migrationsverket/midas/commit/b0e8733489))
+
+## 0.0.12 (2026-03-13)
+
+### 🧱 Updated Dependencies
+
+- Updated components to 17.8.1
+
+## 0.0.11 (2026-03-12)
+
+### 🧱 Updated Dependencies
+
+- Updated components to 17.8.0
+
 ## 0.0.10 (2026-03-12)
 
 ### 🧱 Updated Dependencies