+++ packages/select-styles/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 1.1.37 (2026-03-18)
+
+### 🧱 Updated Dependencies
+
+- Updated theme to 3.12.13
+
+## 1.1.36 (2026-03-17)
+
+This was a version bump only for select-styles to align it with other projects, there were no code changes.
+
+## 1.1.35 (2026-03-16)
+
+### 🧱 Updated Dependencies
+
+- Updated theme to 3.12.12
+
 ## 1.1.34 (2026-03-12)
 
 ### 🧱 Updated Dependencies